Solve issues with the heritability rule
If the covariance matrix has NaNs it currently crashes;
perhaps better to provide a NaN estimate and move on.

A "legitimate" reason for the covariance matrix to have empty spots
are the absence of variability in STs/lineages, which can be caused
by a test dataset or a misconfiguration of mlst/poppunk.

The implemented solution is to "harden" the prepare_fiesta.py
script, and to tollerate crashes for the albi.py script.