Skip the second pass of panfeed for annotate_summary
In some cases, `panfeed-get-kmers` may require a lot of RAM,
which may be inconvenient for some users, especially
since it is not strictly needed for the `annotate_summary`
rule. This change bypasses the second pass of panfeed to complete
the rule, leaving the potentially onerous step to the more explicit
`panfeed` rule.

Tested on an actual Pseudomonas dataset by Marco/Jenny